add all error

+++ libft/ft_atoi.c
@@ -15,7 +15,7 @@
 // #include <stdlib.h>
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	int		i;
 	int		neg;
